pyproject.toml
[project]
name = "cortana-discord-assistant"
version = "0.1.0"
description = "Discord AI assistant bot integrating MemU memory with OpenRouter via PydanticAI"
requires-python = ">=3.10"
dependencies = [
    "discord.py",
    "pydantic",
    "pydantic-ai-slim",
    "pydantic-settings",
    "openai",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
memu = [
    "memu-py",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
